generateTLDs: Fix the chunkSize check
We break the chunks of strings at 65535 bytes because VS 2015 cannot
handle longer strings than this.
However, the logic was wrong and we would not include the length of the
first string in the next chunk, which could lead to a chunk being too
large.

And regenerate the suffix list just in case something went wrong.

Amends 0526ab34af3fdfa2b0741ee621aaa48434067c55.

Change-Id: I004f07913874eca14dfc14fc6c9de8cfe8351692
Reviewed-by: Marc Mutz <marc.mutz@qt.io>