qHash(): add explicit bool overload
Code like in qcborvalue.cpp assumed that worked, and it did, but via
the 1-to-2-arg qHash adapter:

    unsigned long qHash<bool, , true>(bool const&, unsigned long):
        pushq       %rbp
        movq        %rsp, %rbp
        subq        $16, %rsp
        movq        %rdi, -8(%rbp)
        movq        %rsi, -16(%rbp)
        movq        -8(%rbp), %rax
        movzbl      (%rax), %eax
        movzbl      %al, %eax
        movl        $0, %esi
        movl        %eax, %edi
        call        qHash(int, unsigned long)
        xorq        -16(%rbp), %rax
        leave
        ret

Unfortunately, the 1-to-2-arg adapter xors the seed into the result of
the bit-mixing that QHashPrivate::hash() performs, while
QHashPrivate::hash() xors the seed in _before_ mixing, meaning we
can't add the qHash(bool) overload, however well-constrained, in the
same style as the other arithmetic qHash() overloads, because the
results will, in general, differ.

Add the overload in the old style, then, and pre-program the
simplification for Qt 7.

Fixes: QTBUG-126674
Change-Id: I5607b460a499939e1d14886eb5353b01768a87b4
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>