QUuid: make more methods constexpr
Use QT6_{DECL,CALL}_NEW_OVERLOAD mechanism to provide new constexpr
overloads for isNull(), variant(), and version(). Move the old
methods into removed_api.cpp

Use this as an opportunity to optimize the implementaiton of the
methods. Also add compile-time checks to the tests.

It turns out that QNX compiler fails to compile the default
QUuid::isNull() implementation in constexpr context, so rewrite
it using a loop.

Change-Id: Ia23c81194ea27b369f00ce73b016695600187e46
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>