Remove the struct NetlinkSocket
Started with fixing a "Class with destructor should also declare a
copy or move constructor and assignment operator." warning found by an
Axivion scan.

However, the NetlinkSocket struct is used only once and the behavior
can be replicated with a qScopeGuard.

Remove the struct NetlinkSocket and use a qScopeGuard for resource
management.

Drive-by: remove a Q_UNLIKELY that's not needed (qErrnoWarning() is
already marked as Q_DECL_COLD_FUNCTION).

Task-number: QTBUG-125026
Pick-to: 6.8 6.7 6.5
Change-Id: I6b770e284fdda3ff97b8cabac205e9b5abe88b14
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>