Introduce QT_SKIP_DEFAULT_TESTCASE_DIRS target property
The property allows skipping to set the QT_TESTCASE_BUILDDIR/SOURCEDIR
definitions. This might be helpful in certain uscases, to optimize the
build by avoiding the need to set the unused compilation flags.

Add the QT_SKIP_DEFAULT_TESTCASE_DIRS variable that is used as the
default value for the QT_SKIP_DEFAULT_TESTCASE_DIRS property.

Task-number: QTBUG-126729
Change-Id: I6fe5bed428175cdd30d21741c20b745d85f3917d
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>