rhi: gl: Add a helper virtual to punch through to glBufferSubData
The default implementation of the virtual is of course just
memcopying into beginFullDynamicBufferUpdateForCurrentFrame()'s
result, so not very useful for any backend except OpenGL where
(for non-uniform buffers) it is implemented rather with
glBufferSubData (leaving it up to the OpenGL implementation what's
going to happen internally).

The value is somewhat limited in practice, however one user is
going to be Qt Quick: to get as identical as possible results
to Qt 5 in artificial "benchmark" scenes, it becomes important
to go directly to glBufferSubData (just as Qt 5 did) when
updating geometry for a large number of items in every frame,
and skip any intermediate resource update logic.

Pick-to: 6.8
Task-number: QTBUG-125087
Change-Id: I780a1431e021b90590b493e1fb82334cd71bd75b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>